refactor: split high-complexity files by responsibility (drops every qlty smell)
filterutils.go: extract primitive-type coercion helpers (convertToInt,
convertToFloat64, convertToTime, plus dispatchers and string parsers)
into filterconv.go. Same package, identical behavior.

dynamictablefilter/engine.go: extract per-type condition evaluators
(evaluateCondition, evaluateStringCondition, evaluateIntCondition,
evaluateFloat64Condition, evaluateBoolCondition, evaluateTimeCondition,
plus dispatch maps and helpers) into engine_evaluators.go, and the
filter-tree recursion (applyFilterRecursive, isNotFilter, applyNotFilter,
isLeafCondition, applyLeafCondition, applyGroupFilter, foldGroupFilter,
combineLogicalMatch, evaluateGroupStep) into engine_recursion.go.
Replaces inline switch in evaluateBoolCondition with a boolComparators
map for symmetry with numericComparators.

main.go: extract HTTP handlers (filterHandler, decodeFilterRequest,
runFilterQuery, runTransactionQuery, listFilterableEntitiesHandler,
listDynamicTablesHandler, dynamicTablesItemHandler,
dynamicTableSchemaHandler, dynamicTableFilterHandler, spaFallbackHandler,
plus filterRequestBody and apiPathPrefixes) into main_handlers.go, and
seed-data generators (generateTransactions, generateTest1SchemaData,
generateTest2SchemaData, generateTest3SchemaData) into main_seed.go.

transaction_filter_test.go: split TestFilterTransactions
(cyclomatic complexity 42) into runFilterTransactionTestCase plus
helpers (simulateFilterError via simulatedErrors map,
reportFilterErrorOutcome, reportFilterCountMismatch). Lifts the test
case slice into filterTransactionTestCases().

Build + tests pass. qlty smells --all now reports zero issues for
this repo (was: 4 smells before this change).
