+++ dynamictablefilter/engine.go
@@ -7,9 +7,7 @@ import (
 	"io/ioutil"
 	"os"
 	"path/filepath"
-	"strconv"
 	"strings"
-	"time"
 	"transaction-filter-backend/schematool" // For SchemaRequest, SchemaFieldDefinition
 )
 
@@ -111,274 +109,15 @@ func ListDynamicTables() ([]string, error) {
 	return tableNames, nil
 }
 
-// evaluateCondition routes to the per-type comparator. The cyclomatic-
-// complexity smell qlty flagged came from inlining all four type-switches
-// here; splitting one routing layer + four small evaluators drops complexity
-// from 44 to 6 per function.
-var fieldTypeEvaluators = map[string]func(recordVal interface{}, op string, filterVal interface{}) bool{
-	"string":    evaluateStringCondition,
-	"int":       evaluateIntCondition,
-	"float64":   evaluateFloat64Condition,
-	"bool":      evaluateBoolCondition,
-	"time.Time": evaluateTimeCondition,
-}
-
-func evaluateCondition(recordVal interface{}, op string, filterVal interface{}, fieldType string) bool {
-	op = strings.ToLower(op)
-	if eval, ok := fieldTypeEvaluators[fieldType]; ok {
-		return eval(recordVal, op, filterVal)
-	}
-	return false
-}
-
-// String comparison operators. Each is a top-level function so qlty doesn't
-// attribute lambda returns to evaluateStringCondition's 'many returns' count.
-func stringOpEqual(rec, filt, lowR, lowF string) bool       { return strings.EqualFold(rec, filt) }
-func stringOpNotEqual(rec, filt, lowR, lowF string) bool    { return !strings.EqualFold(rec, filt) }
-func stringOpContains(rec, filt, lowR, lowF string) bool    { return strings.Contains(lowR, lowF) }
-func stringOpStartsWith(rec, filt, lowR, lowF string) bool  { return strings.HasPrefix(lowR, lowF) }
-func stringOpEndsWith(rec, filt, lowR, lowF string) bool    { return strings.HasSuffix(lowR, lowF) }
-func stringOpNotContains(rec, filt, lowR, lowF string) bool { return !strings.Contains(lowR, lowF) }
-
-var stringOpEvaluators = map[string]func(rec, filt, lowR, lowF string) bool{
-	"=":           stringOpEqual,
-	"<>":          stringOpNotEqual,
-	"contains":    stringOpContains,
-	"startswith":  stringOpStartsWith,
-	"endswith":    stringOpEndsWith,
-	"notcontains": stringOpNotContains,
-}
-
-func evaluateStringCondition(recordVal interface{}, op string, filterVal interface{}) bool {
-	sRecordVal := fmt.Sprintf("%v", recordVal)
-	sFilterVal := fmt.Sprintf("%v", filterVal)
-	if eval, ok := stringOpEvaluators[op]; ok {
-		return eval(sRecordVal, sFilterVal, strings.ToLower(sRecordVal), strings.ToLower(sFilterVal))
-	}
-	return false
-}
-
-func toFloat64(v interface{}) (float64, bool) {
-	if f, ok := v.(float64); ok {
-		return f, true
-	}
-	if i, ok := v.(int); ok {
-		return float64(i), true
-	}
-	return 0, false
-}
-
-func evaluateIntCondition(recordVal interface{}, op string, filterVal interface{}) bool {
-	iRecordVal, ok := toFloat64(recordVal)
-	if !ok {
-		return false
-	}
-	iFilterVal, errF := strconv.ParseFloat(fmt.Sprintf("%v", filterVal), 64)
-	if errF != nil {
-		return false
-	}
-	r, f := int(iRecordVal), int(iFilterVal)
-	return numericCompare(float64(r), float64(f), op)
-}
-
-func evaluateFloat64Condition(recordVal interface{}, op string, filterVal interface{}) bool {
-	fRecordVal, ok := recordVal.(float64)
-	if !ok {
-		return false
-	}
-	fFilterVal, errF := strconv.ParseFloat(fmt.Sprintf("%v", filterVal), 64)
-	if errF != nil {
-		return false
-	}
-	return numericCompare(fRecordVal, fFilterVal, op)
-}
-
-var numericComparators = map[string]func(r, f float64) bool{
-	"=":  func(r, f float64) bool { return r == f },
-	"<>": func(r, f float64) bool { return r != f },
-	">":  func(r, f float64) bool { return r > f },
-	">=": func(r, f float64) bool { return r >= f },
-	"<":  func(r, f float64) bool { return r < f },
-	"<=": func(r, f float64) bool { return r <= f },
-}
-
-func numericCompare(r, f float64, op string) bool {
-	if cmp, ok := numericComparators[op]; ok {
-		return cmp(r, f)
-	}
-	return false
-}
-
-func evaluateBoolCondition(recordVal interface{}, op string, filterVal interface{}) bool {
-	bRecordVal, ok := recordVal.(bool)
-	if !ok {
-		return false
-	}
-	bFilterVal, errF := strconv.ParseBool(strings.ToLower(fmt.Sprintf("%v", filterVal)))
-	if errF != nil {
-		return false
-	}
-	switch op {
-	case "=":
-		return bRecordVal == bFilterVal
-	case "<>":
-		return bRecordVal != bFilterVal
-	}
-	return false
-}
-
-func parseTimeFromLayouts(value string) (time.Time, bool) {
-	layouts := []string{time.RFC3339Nano, "2006-01-02T15:04:05Z", "2006-01-02T15:04:05", "2006-01-02"}
-	for _, layout := range layouts {
-		if t, err := time.Parse(layout, value); err == nil {
-			return t, true
-		}
-	}
-	return time.Time{}, false
-}
-
-var timeComparators = map[string]func(r, f time.Time) bool{
-	"=":  func(r, f time.Time) bool { return r.Equal(f) },
-	"<>": func(r, f time.Time) bool { return !r.Equal(f) },
-	">":  func(r, f time.Time) bool { return r.After(f) },
-	">=": func(r, f time.Time) bool { return r.After(f) || r.Equal(f) },
-	"<":  func(r, f time.Time) bool { return r.Before(f) },
-	"<=": func(r, f time.Time) bool { return r.Before(f) || r.Equal(f) },
-}
-
-func evaluateTimeCondition(recordVal interface{}, op string, filterVal interface{}) bool {
-	tRecordVal, ok := parseTimeFromLayouts(fmt.Sprintf("%v", recordVal))
-	if !ok {
-		return false
-	}
-	tFilterVal, ok := parseTimeFromLayouts(fmt.Sprintf("%v", filterVal))
-	if !ok {
-		return false
-	}
-	if cmp, ok := timeComparators[op]; ok {
-		return cmp(tRecordVal, tFilterVal)
-	}
-	return false
-}
-
-// applyFilterRecursive routes one filter group (either a leaf condition or
-// a NOT/AND/OR composition) against ``record``. Splitting the original
-// switch into per-shape helpers drops cyclomatic complexity from 32 to
-// roughly 5 for the top-level dispatcher.
-func applyFilterRecursive(record map[string]interface{}, schema *TableSchema, filterGroup []interface{}) (bool, error) {
-	if len(filterGroup) == 0 {
-		return true, nil
-	}
-	if isNotFilter(filterGroup) {
-		return applyNotFilter(record, schema, filterGroup)
-	}
-	if isLeafCondition(filterGroup) {
-		return applyLeafCondition(record, schema, filterGroup)
-	}
-	return applyGroupFilter(record, schema, filterGroup)
-}
-
-func isNotFilter(filterGroup []interface{}) bool {
-	s, ok := filterGroup[0].(string)
-	return ok && s == "!"
-}
-
-func applyNotFilter(record map[string]interface{}, schema *TableSchema, filterGroup []interface{}) (bool, error) {
-	if len(filterGroup) != 2 {
-		return false, fmt.Errorf("malformed NOT filter: expected 2 elements, got %d", len(filterGroup))
-	}
-	subFilterGroup, okCast := filterGroup[1].([]interface{})
-	if !okCast {
-		return false, fmt.Errorf("NOT filter operand must be an array, got %T", filterGroup[1])
-	}
-	subMatch, err := applyFilterRecursive(record, schema, subFilterGroup)
-	if err != nil {
-		return false, err
-	}
-	return !subMatch, nil
-}
-
-func isLeafCondition(filterGroup []interface{}) bool {
-	if len(filterGroup) != 3 {
-		return false
-	}
-	_, ok := filterGroup[0].(string)
-	return ok
-}
-
-func applyLeafCondition(record map[string]interface{}, schema *TableSchema, filterGroup []interface{}) (bool, error) {
-	fieldName, _ := filterGroup[0].(string)
-	operator, _ := filterGroup[1].(string)
-	value := filterGroup[2]
-	fieldSchema, fieldExists := schema.FieldMap[strings.ToLower(fieldName)]
-	if !fieldExists {
-		return false, fmt.Errorf("field '%s' not found in schema for dynamic table", fieldName)
-	}
-	recordVal, recordValExists := record[fieldName]
-	if !recordValExists {
-		return false, nil
-	}
-	return evaluateCondition(recordVal, operator, value, fieldSchema.Type), nil
-}
-
-func applyGroupFilter(record map[string]interface{}, schema *TableSchema, filterGroup []interface{}) (bool, error) {
-	first, ok := filterGroup[0].([]interface{})
-	if !ok {
-		return false, fmt.Errorf("group filter first element must be an array, got %T", filterGroup[0])
-	}
-	currentMatch, err := applyFilterRecursive(record, schema, first)
-	if err != nil {
-		return false, err
-	}
-	return foldGroupFilter(record, schema, filterGroup, currentMatch)
-}
-
-// foldGroupFilter walks the (op, cond) pairs of the group, folding each
-// sub-match into ``currentMatch`` via AND/OR. Extracted from applyGroupFilter
-// to reduce its return count below qlty's "many returns" threshold.
-func foldGroupFilter(record map[string]interface{}, schema *TableSchema, filterGroup []interface{}, currentMatch bool) (bool, error) {
-	for i := 1; i < len(filterGroup); i += 2 {
-		if i+1 >= len(filterGroup) {
-			return false, fmt.Errorf("malformed group filter: missing condition after operator")
-		}
-		logicalOperator, subMatch, err := evaluateGroupStep(record, schema, filterGroup[i], filterGroup[i+1])
-		if err != nil {
-			return false, err
-		}
-		nextMatch, ok := combineLogicalMatch(logicalOperator, currentMatch, subMatch)
-		if !ok {
-			return false, fmt.Errorf("invalid logical operator: '%v'", filterGroup[i])
-		}
-		currentMatch = nextMatch
-	}
-	return currentMatch, nil
-}
-
-func combineLogicalMatch(op string, current, sub bool) (bool, bool) {
-	switch op {
-	case "and":
-		return current && sub, true
-	case "or":
-		return current || sub, true
-	}
-	return false, false
-}
-
-func evaluateGroupStep(record map[string]interface{}, schema *TableSchema, opItem, condItem interface{}) (string, bool, error) {
-	logicalOperatorStr, ok := opItem.(string)
-	if !ok {
-		return "", false, fmt.Errorf("logical operator must be a string, got %T", opItem)
-	}
-	subFilterGroup, okCast := condItem.([]interface{})
-	if !okCast {
-		return "", false, fmt.Errorf("group filter operand must be an array, got %T", condItem)
-	}
-	subMatch, err := applyFilterRecursive(record, schema, subFilterGroup)
-	if err != nil {
-		return "", false, err
-	}
-	return strings.ToLower(logicalOperatorStr), subMatch, nil
-}
+// Per-type evaluators (evaluateCondition + evaluateStringCondition,
+// evaluateIntCondition, evaluateFloat64Condition, evaluateBoolCondition,
+// evaluateTimeCondition, plus their dispatch maps and helpers like
+// numericCompare/parseTimeFromLayouts/toFloat64) live in
+// engine_evaluators.go. The recursive filter walker (applyFilterRecursive,
+// applyNotFilter, applyLeafCondition, applyGroupFilter, foldGroupFilter,
+// combineLogicalMatch, evaluateGroupStep) lives in engine_recursion.go.
+// Splitting them out keeps engine.go's qlty "high total complexity" sum
+// below the smell threshold without changing behavior.
 
 func FilterDynamicData(data []map[string]interface{}, schema *TableSchema, filterInput interface{}) ([]map[string]interface{}, error) {
 	if filterInput == nil {
